Merge ios-init to develop

+++ Paco-iOS/Paco/source/core/common/PacoUtility.h
@@ -18,12 +18,12 @@
 @interface PacoUtility : NSObject
 
 //YMZ:TODO: should consider if we need to change max to be exclusive
-+ (NSUInteger)randomUnsignedIntegerBetweenMin:(NSUInteger)min andMax:(NSUInteger)max;
++ (unsigned int)randomUnsignedIntegerBetweenMin:(unsigned int)min andMax:(unsigned int)max;
 
 //YMZ:TODO: should check if we should allow duplicate numbers
-+ (NSArray*)randomIntegersInRange:(NSUInteger)rangeNumber
-                    numOfIntegers:(NSUInteger)numOfIntegers
-                        minBuffer:(NSUInteger)minBuffer;
++ (NSArray*)randomIntegersInRange:(unsigned int)rangeNumber
+                    numOfIntegers:(int)numOfIntegers
+                        minBuffer:(int)minBuffer;
 
 
 @end
